[refactor] Customer 관련 생성 완료
