+++ client/src/Navigator/MainTab/Customer/bidTabScreen.js
@@ -3,13 +3,19 @@ import { Stack } from '../../../../App';
 import { createStackNavigator } from '@react-navigation/stack';
 import CommonHeader from '../../../Components/HeaderBar/commonHeader';
 
-import CheckingMatchingScreen from '../../../Screen/02_SecondTab/User/checkingMatchingScreen';
+import CheckingMatchingScreen from '../../../Screen/02_SecondTab/Customer/checkingMatchingScreen';
 
-import UpdateProposalScreen from '../../../Screen/02_SecondTab/User/waitMatching/updateProposalScreen';
-import UpdateAfterImageScreen from '../../../Screen/02_SecondTab/User/waitMatching/updateAfterImageScreen';
-import UpdateFromAlbumScreen from '../../../Screen/02_SecondTab/User/waitMatching/updateAfterImage/updateFromAlbumScreen';
-import UpdateFromBidiScreen from '../../../Screen/02_SecondTab/User/waitMatching/updateAfterImage/updateFromBidiScreen';
-import UpdateFromScrapScreen from '../../../Screen/02_SecondTab/User/waitMatching/updateAfterImage/updateFromScrapScreen';
+import CreateProposalScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/createProposalScreen';
+import SelectAfterImageScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/selectAfterImageScreen';
+import SelectFromAlbumScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/selectFromAlbumScreen';
+import SelectFromBidiScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/selectFromBidiScreen';
+import SelectFromScrapScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/selectFromScrapScreen';
+
+import UpdateProposalScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/updateProposalScreen';
+import UpdateAfterImageScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateAfterImageScreen';
+import UpdateFromAlbumScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateFromAlbumScreen';
+import UpdateFromBidiScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateFromBidiScreen';
+import UpdateFromScrapScreen from '../../../Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateFromScrapScreen';
 
 const bidStack = createStackNavigator();
 
@@ -18,8 +24,33 @@ function BidStackScreen() {
     <Stack.Navigator>
       <bidStack.Screen name="check" component={CheckingMatchingScreen} options={CommonHeader} />
       <bidStack.Screen
+        name="CreateProposal"
+        component={CreateProposalScreen}
+        options={CommonHeader}
+      />
+      <bidStack.Screen
         name="UpdateProposal"
         component={UpdateProposalScreen}
+        options={CommonHeader}
+      />
+      <bidStack.Screen
+        name="SelectAfterImage"
+        component={SelectAfterImageScreen}
+        options={CommonHeader}
+      />
+      <bidStack.Screen
+        name="SelectFromAlbum"
+        component={SelectFromAlbumScreen}
+        options={CommonHeader}
+      />
+      <bidStack.Screen
+        name="SelectFromScrap"
+        component={SelectFromScrapScreen}
+        options={CommonHeader}
+      />
+      <bidStack.Screen
+        name="SelectFromBidi"
+        component={SelectFromBidiScreen}
         options={CommonHeader}
       />
       <bidStack.Screen
