+++ client/src/Screen/02_SecondTab/Customer/checkingMatchingScreen.js
@@ -21,7 +21,6 @@ function CheckingMatchingScreen({ navigation }) {
   useEffect(() => {
     dispatch(getMatchingByCustomerId(user.id));
   }, [dispatch]);
-  console.log(matching, loading, error);
   if (loading || error || !matching) return <Loading loading />;
   return (
     <View style={{ flex: 1 }}>
