+++ client/src/Screen/02_SecondTab/Customer/waitMatching/TabScreen/myProposalScreen.js
@@ -12,14 +12,14 @@ import {
 import { useSelector, useDispatch } from 'react-redux';
 
 // Components
-import ProposalUserInfo from '../../../../Components/Proposal/proposalUserInfo';
-import BottomButton from '../../../../Components/Common/bottomButton';
+import ProposalUserInfo from '../../../../../Components/Proposal/proposalUserInfo';
+import BottomButton from '../../../../../Components/Common/bottomButton';
 
 // API
-import ProposalAPI from '../../../../Api/proposal';
+import ProposalAPI from '../../../../../Api/proposal';
 
 // Redux Action
-import { deleteProposal } from '../../../../Contexts/Proposal/action';
+import { deleteProposal } from '../../../../../Contexts/Proposal/action';
 
 function MyProposalScreen({ navigation, proposal }) {
   const [imageToggle, setImageToggle] = useState(false);
