+++ client/src/Screen/02_SecondTab/Customer/waitMatching/createProposalScreen.js
@@ -14,17 +14,13 @@ import {
 
 // Components
 import DropDownPicker from 'react-native-dropdown-picker';
-import { KEYWORDS } from '../../../Lib/constant';
+import { KEYWORDS } from '../../../../Lib/constant';
 import { LogBox } from 'react-native';
 LogBox.ignoreLogs(['Non-serializable values were found in the navigation state']);
-import BottomButton from '../../../Components/Common/bottomButton';
+import BottomButton from '../../../../Components/Common/bottomButton';
 
 // API
-import ProposalAPI from '../../../Api/proposal';
-
-// Redux Action
-import { registerWithFile, registerProposal } from '../../../Contexts/Proposal/action';
-import { registerMatching } from '../../../Contexts/Matching/action';
+import ProposalAPI from '../../../../Api/proposal';
 
 function CreateProposalScreen({ navigation }) {
   const { data: user } = useSelector((state) => state.user);
@@ -127,13 +123,13 @@ function CreateProposalScreen({ navigation }) {
       };
       if (isFromAlbum) {
         const proposal = await ProposalAPI.registerWithFile(createFormData(albumImage, body));
-        navigation.replace('Main');
+        navigation.replace('check');
       } else {
         const proposal = await ProposalAPI.registerProposal({
           ...body,
           after_src: afterImageStyle,
         });
-        navigation.replace('Main');
+        navigation.replace('check');
       }
     }
   };
