+++ client/src/Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/introStyleScrapScreen.js
