+++ client/src/Screen/02_SecondTab/Customer/waitMatching/selectAfterImage/selectFromBidiScreen.js
@@ -4,7 +4,7 @@ import { Text, View, Image, ScrollView, StyleSheet, TouchableOpacity } from 'rea
 
 // Components
 import Ionicons from 'react-native-vector-icons/Ionicons';
-import { STYLE_INFO } from '../../../../Lib/constant';
+import { STYLE_INFO } from '../../../../../Lib/constant';
 
 function SelectFromBidiScreen({ navigation, route }) {
   const { setAfterImageStyle } = route.params;
@@ -19,7 +19,7 @@ function SelectFromBidiScreen({ navigation, route }) {
     setAfterImageStyle(
       `https://bidi-s3.s3.ap-northeast-2.amazonaws.com/image/user/${user.id}/result/${afterStyle}.jpg`,
     );
-    navigation.navigate('Main');
+    navigation.navigate('CreateProposal');
   };
   const selectStyle = (style) => {
     setAfterStyle(style);
