+++ client/src/Screen/02_SecondTab/Customer/waitMatching/staticScreen/noBidScreen.js
