+++ client/src/Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateFromBidiScreen.js
