+++ client/src/Screen/02_SecondTab/Customer/waitMatching/updateAfterImage/updateFromScrapScreen.js
